updated the new chapter form for the last time.
readded the route, button, and created form on a seperate component from the chapter info
